+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 - Update `contribution` and `contributionAndFinalize` detector tests.
   TEE tasks with callback are now eligible to `contributeAndFinalize` flow. (#719)
 - Resolve deprecations caused by `TaskDescription` in `ReplicatesService` and `ResultService`. (#723)
+- Add missing `@PreDestroy` annotation to `purgeAllTasksData` methods in `IexecHubService` and `ReplicateSupplyService`. (#724)
 
 ### Dependency Upgrades
 
@@ -142,7 +143,8 @@ All notable changes to this project will be documented in this file.
 - Replace single thread executor with synchronized keyword. (#633)
 - Move contribution status checks from `iexec-commons-poco`. (#636)
 - Use `BlockchainAdapterService` from `iexec-blockchain-adapter-api-library`. (#641)
-- `ResultRepositoryConfiguration` and `WorkerConfiguration` classes are now immutable with `@Value` lombok annotation. (#650)
+- `ResultRepositoryConfiguration` and `WorkerConfiguration` classes are now immutable with `@Value` lombok annotation. (
+  #650)
 
 ### Bug Fixes
 
@@ -153,7 +155,8 @@ All notable changes to this project will be documented in this file.
 - Use less MongoDB calls when updating a task to a final status. (#649)
 - Save contribution and result updload replicate data when `CONTRIBUTE_AND_FINALIZE_DONE`. (#651)
 - Fix potential `NullPointerException` during first worker replicate request. (#652)
-- Fix missed replicate status update detectors to avoid false positives by mixing `CONTRIBUTE-REVEAL-FINALIZE` and `CONTRIBUTE_AND_FINALIZE` workflows. (#653)
+- Fix missed replicate status update detectors to avoid false positives by mixing `CONTRIBUTE-REVEAL-FINALIZE` and
+  `CONTRIBUTE_AND_FINALIZE` workflows. (#653)
 
 ### Dependency Upgrades