+++ CHANGELOG.md
@@ -30,6 +30,11 @@ All notable changes to this project will be documented in this file.
 - Upgrade to Gradle 8.10.2. (#714)
 - Upgrade to `testcontainers` 1.20.4. (#721)
 - Upgrade to `mongo:7.0.15-jammy`. (#722)
+- Upgrade to `iexec-commons-poco` 4.2.0. (#725)
+- Upgrade to `iexec-common` 8.6.0. (#725)
+- Upgrade to `iexec-blockchain-adapter-api-library` 8.6.0. (#725)
+- Upgrade to `iexec-result-proxy-library` 8.6.0. (#725)
+- Upgrade to `iexec-sms-library` 8.7.0. (#725)
 
 ## [[8.5.0]](https://github.com/iExecBlockchainComputing/iexec-core/releases/tag/v8.5.0) 2024-06-19